UpdateConfFile: implement 'remove' option
By default, if a file is removed in the commit, nothing happens to its
copy in the repository. If one wants to have the copy removed from the
repository when the file is deleted, there's a new option 'remove'
that should be set to true.

This feature was suggested by Uwe Satthoff. Thanks!

Signed-off-by: Gustavo L. de M. Chaves <gustavo@cpqd.com.br>