Make hooks execute in the order they are defined
The hooks defined via the hook directives were kept in a hash keyed by
the hook routine's address. Since they were executed as returned by the
'value' hash function the order wasn't really specified.

This commit substitute a list for that hash and makes sure that the
hooks are executed in the order they're defined.