UpdateConfFile: fix detection of out of conf destination
We tried to detect if the destination of an updated conf file was
outside the repository's conf directory by using the Cwd::abs_path
function. However, since the destination could well be a path missing
the innermost directory components the Cwd::abs_path returned undef in
this cases, which was wrongly understood as meaning that the destination
was outside the conf directory.

Now we use another approach. We simply check if the TO argument of the
UPDATE_CONF_FILE directive (after being subject to interpolation if it's
really a Regexp) contains a '..' path component. If so, the hook
complains and doesn't generate the updated conf file.