Avoid repeated invocation of the same hooks
Commit 68037717 tried to guarantee that the hooks were executed in the
order they were configured by the plugin directives. To do that we
substituted an array of functions (@SVN::Hooks::Hooks) for the hash
(%SVN::Hooks::Hooks) that was used to hold the function
references. This data structure was used by the hook directives (such
as SVN::Hooks::PRE_COMMIT) to "hook" the functions to be executed
later.

However, the implementation introduced a bug so that each invocation
of a hook directive made the hook be invoked yet again. So, if, for
example, someone used 10 CHECK_JIRA directives the pre-commit hook
would be invoked 10 times!

This commit tries to do it right by keeping a hash and an array. The
hash is used to keep track of which functions were already hooked and
the array is used to keep their order.