UpdateConfFile: avoid filetests
The -x filetest does not work as I expected. It seems to actually check
if the file is owned by the EUID and if it has the executable bit set
for the owner. If the file belongs to another user it doesn't matter if
if's executable by the users EGID or by others.

So, we don't check that upon reading the configuration and let it try to
invoke the command later on. Hopefully, if the invocation fails, the
error message will be sufficiently clear.